chore: cleanup fixtures
